Change cutout to first ask for a small section and then the complete cutout.
Sometimes cutouts cross the boundary between extensions.  Doing a small section first and then using that to get
the name of the extension that should be used for the larger cutout is a bit of a hack to allow us to determine which
extension should be the one that contains the object.   Now we only keep that Single Extension... Also, cast that to
a PrimaryHDU for writing since we want not have Extensions.... ever.